remove dummy mp3s

+++ spider/fileUtil.py
@@ -5,14 +5,17 @@
 
 __author__ = 'XSunny'
 
+import os
+
 #文件工具 - 包含写字节文件和字符文件，并且可以存取map对象
 
 def saveByteFile(fileName, data):
 	#print fileName
 	f = None
 	try:
-		f = open(fileName, "wb")
-		f.write(data)
+		if not os.path.isfile(fileName):
+			f = open(fileName, "wb")
+			f.write(data)
 	except IOError, e:
 		print "FileIO ERROR!"
 	finally:
@@ -22,8 +25,9 @@ def saveByteFile(fileName, data):
 def saveTextFile(fileName, data):
 	f = None
 	try:
-		f = open(fileName, "w")
-		f.write(data)
+		if not os.path.isfile(fileName):
+			f = open(fileName, "w")
+			f.write(data)
 	except IOError, e:
 		print "FileIO ERROR!"
 	finally:
